Reuse dumptruck-web again
sql.js unsuitable for large databases.

Only dependency on a box is readSettings; for a standalone app,
this path could be hardcoded, but need to figure out how this
should be configured first.

Still uses scraperwiki.alert; this potentially could be moved out
of scraperwiki.js to make it completely independent of
scraperwiki.js.